[#6]/refactor SectionTitle
