+++ landing/src/assets/styles/utils.ts
@@ -10,3 +10,9 @@ export const flexColumn = css`
   display: flex;
   flex-direction: column;
 `;
+
+export const flexCenter = css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
